requirements.txt
# Hand-curated direct dependencies. For exact reproduction of the
# environment used during development see requirements.lock.txt.

# ── Core data ──────────────────────────────────────────────────────────
pandas==2.3.3
numpy==2.0.2
openpyxl==3.1.5
scipy==1.13.1

# ── ML / Statistics ────────────────────────────────────────────────────
scikit-learn==1.6.1
xgboost==2.1.4
statsmodels==0.14.6

# ── Deep Learning (CPU build is fine; PyPI default works) ──────────────
torch==2.8.0

# ── Forecasting ────────────────────────────────────────────────────────
prophet==1.3.0
cmdstanpy==1.3.0

# ── Feature engineering ────────────────────────────────────────────────
holidays==0.83

# ── API ────────────────────────────────────────────────────────────────
fastapi==0.128.8
uvicorn[standard]==0.39.0
pydantic==2.13.3
python-dotenv==1.2.1

# ── Utilities ──────────────────────────────────────────────────────────
pyyaml==6.0.3
joblib==1.5.3
matplotlib==3.9.4

# ── Tests ──────────────────────────────────────────────────────────────
pytest==8.4.2
httpx==0.28.1
